Revised Intro
In order to understand the mechanisms of aging as accurately as
possible, the organism Saccharomyces Cervisiae was observed because of
its similar complex internal cell structure to higher-level eukaryotes
In yeast Saccharomyces Cerevisiae, aging is known as replicative life
span or RLS; replicative life span is determined by the number of cell
divisions that occur prior to senescence (when diploid cells can no
longer divide). I studied the effect of robustness on RLS in S.
cerevisiae. Previous research has provided evidence that cellular aging
is an emergent property of gene networks. These gene networks allow the
cell to adapt and survive, and thus depicts the robustness of the cell.
As a cell’s network robustness decreases, it will be less able to adapt
against external perturbations. As a result, functionality of protein
activities are depleted; Aging is basically the increase of chance of
dying with age.
  Protein robustness, the ability of a cell to maintain homeostasis,
was found to have a positive correlation with RLS, indicating an
increased lifespan in prior studies. It is hypothesized that robustness
and several different proxies will be correlated to replicative life
span in Saccharomyces Cerevisiae. The robustness proxies examined are
morphological plasticity robustness and gene expression plasticity.
These proxies are a way to measure the robustness of a system.